Revert "image uploader toolbar for post/page editor"
