+++ package.json
@@ -31,7 +31,7 @@
     "@tailwindcss/typography": "^0.5.3",
     "@tippy.js/react": "^3.1.1",
     "@typescript-eslint/eslint-plugin": "^5.38.0",
-    "@uiw/react-md-editor": "^3.20.0",
+    "@uiw/react-md-editor": "^3.19.7",
     "apollo-server-micro": "^3.10.2",
     "babel-plugin-transform-remove-imports": "^1.7.0",
     "cheerio": "^1.0.0-rc.10",
