+++ src/components/Writing/Editor/PostEditorComposer.tsx
@@ -4,7 +4,7 @@ import * as React from 'react'
 import { Dropzone } from '~/components/Dropzone'
 import { Textarea } from '~/components/Input'
 import { Detail } from '~/components/ListDetail/Detail'
-import { CustomizedMDEditor, MDEditor } from '~/components/ReactMdEditor'
+import { MDEditor } from '~/components/ReactMdEditor'
 
 import { PostEditorContext } from './PostEditor'
 
@@ -59,10 +59,14 @@ export function PostEditorComposer({ site }) {
             placeholder={'Post title'}
             className="block w-full p-0 text-2xl font-bold border-none composer text-primary focus:border-0 focus:outline-none focus:ring-0 dark:bg-black md:text-3xl"
           />
-          <CustomizedMDEditor
-            value={draftState.text}
-            onChange={handleTextChange}
-          />
+          <div className="block w-full relative p-0 text-lg font-normal">
+            <MDEditor
+              value={draftState.text}
+              onChange={handleTextChange}
+              preview="edit"
+              height={500}
+            />
+          </div>
           {/* <Textarea
             rows={20}
             maxRows={2000}
